+++ lib/persistence/entities/profession.dart
@@ -1,8 +1,5 @@
-import 'package:battle_it_out/persistence/dao/profession_dao.dart';
-import 'package:battle_it_out/persistence/entities/attribute.dart';
 import 'package:battle_it_out/persistence/entities/dto.dart';
 import 'package:battle_it_out/persistence/entities/skill.dart';
-import 'package:battle_it_out/persistence/wfrp_database.dart';
 
 class Profession extends DTO {
   int id;
